Refs #10807: refactor to move implementations to DomainParticipantImpl
